[feat/#38]: 컴포넌트 분리
